Add minimal KNIME workflow with container input
This adds a workflow containing a Container Input node to retrieve tabular input from the caller.
It allows to hand over tabular data instead of just workflow variables (as with KnimeWrapperTaskBase)

For usage with knimepy, the workflow has to be in a folder structure, calling as .knwf doesn't seem to work.